moved moist_cv test code into translate tests
