+++ fv3core/tests/savepoint/translate/translate_moistcvpluspt_2d.py
@@ -1,12 +1,67 @@
 import fv3core.stencils.moist_cv as moist_cv
+from pace.dsl.dace.orchestrate import computepath_method
+from pace.dsl.stencil import StencilFactory
+from pace.dsl.typing import FloatField
 from pace.stencils.testing import TranslateFortranData2Py, pad_field_in_j
+
+
+class MoistPT:
+    """
+    Class to test with DaCe orchestration. test class is MoistCVPlusPt_2d
+    """
+
+    def __init__(
+        self,
+        stencil_factory: StencilFactory,
+        grid,
+    ):
+        self._moist_cv_pt = stencil_factory.from_origin_domain(
+            moist_cv.moist_pt,
+            origin=grid.compute_origin(),
+            domain=(grid.nic, 1, grid.npz),
+        )
+
+    @computepath_method
+    def __call__(
+        self,
+        qvapor: FloatField,
+        qliquid: FloatField,
+        qrain: FloatField,
+        qsnow: FloatField,
+        qice: FloatField,
+        qgraupel: FloatField,
+        q_con: FloatField,
+        gz: FloatField,
+        cvm: FloatField,
+        pt: FloatField,
+        cappa: FloatField,
+        delp: FloatField,
+        delz: FloatField,
+        r_vir: float,
+    ):
+        self._moist_cv_pt(
+            qvapor,
+            qliquid,
+            qrain,
+            qsnow,
+            qice,
+            qgraupel,
+            q_con,
+            gz,
+            cvm,
+            pt,
+            cappa,
+            delp,
+            delz,
+            r_vir,
+        )
 
 
 class TranslateMoistCVPlusPt_2d(TranslateFortranData2Py):
     def __init__(self, grid, namelist, stencil_factory):
         super().__init__(grid, namelist, stencil_factory)
         self.stencil_factory = stencil_factory
-        self.compute_func = moist_cv.MoistPT(stencil_factory, self.grid)
+        self.compute_func = MoistPT(stencil_factory, self.grid)
         self.in_vars["data_vars"] = {
             "qvapor": {"serialname": "qvapor_js"},
             "qliquid": {"serialname": "qliquid_js"},
